I know it is a horrid mess:
Updates: Attempting to OOP with creation of a list of TaskObjects to give ToDoList persisting data.

Commented code to test if it works. Comments on adding a UI. I want it to be cross-platform instead of Windows Forms specific. I found GTKMM for that, but I recently stumbled across Mongoose and will probably be using that.

Not shown: a failed attempt to print ToDoList. My current focus is testing if the list works. Once this is verified all the test code will be removed and I will work on the UI.